$inject was getting printed as ‘inject’ and becoming a placeholder.
The documentation states that ‘$[value]’ is a snippet variable and if such a variable cannot be found, it defaults to the name of the variable and becomes a placeholder. We typically don’t want to change inject to something else in this case thus removing the need for it to be a placeholder. Escaping the ‘$’ allows us to skip the placeholder and print out the dollar sign.