Fixing inject placeholder and use strict tabs (#14)
* $inject was getting printed as ‘inject’ and becoming a placeholder.

The documentation states that ‘$[value]’ is a snippet variable and if such a variable cannot be found, it defaults to the name of the variable and becomes a placeholder. We typically don’t want to change inject to something else in this case thus removing the need for it to be a placeholder. Escaping the ‘$’ allows us to skip the placeholder and print out the dollar sign.

* Fixing ‘onDestory’ typo and escaping component lifecycle hooks.

* Consistent use of tabs in ‘use strict’;

Cleaning up some of the extra functions and comment lines to make it cleaner.

Placeholder for choosing directive ‘restrict’.

Updating filter to use Filter naming conventions as Angular automatically adds ‘Filter’ as a suffix for each filter registered.

* Removing accidental spacing added in by formatting.

* More spacing…

* The documentation states that ‘$[value]’ is a snippet variable and if such a variable cannot be found, it defaults to the name of the variable and becomes a placeholder. We typically don’t want to change inject to something else in this case thus removing the need for it to be a placeholder. Escaping the ‘$’ allows us to skip the placeholder and print out the dollar sign.

Fixed lifecycle hooks typo: $Destory

Consistent use of tabs in ‘use strict’;